+++ TimJeffPizza/TimJeffPizza/Model/PriceModel.swift
@@ -1,0 +1,14 @@
+//
+//  PriceModel.swift
+//  TimJeffPizza
+//
+//  Created by Obeisun Timothy on 29/10/2021.
+//
+
+import Foundation
+
+// MARK: - Price
+struct Price: Codable {
+    let size: String
+    let price: Double
+}
