+++ TimJeffPizza/TimJeffPizza/View/OnBoardingViewController/OnBoardingViewController.swift
@@ -34,14 +34,3 @@ class OnBoardingViewController: UIViewController {
         setupPresentMainTabBarController(userStatus: false, logInStatus: true)
     }
 }
-
-extension OnBoardingViewController {
-    func setupPresentMainTabBarController(userStatus: Bool, logInStatus: Bool) {
-        UserDefaults.standard.set(userStatus, forKey: "isMarried")
-        UserDefaults.standard.set(logInStatus, forKey: "isLoggedIn")
-        let storyboard = UIStoryboard(name: "Main", bundle: nil)
-        let mainTabBarController = storyboard.instantiateViewController(identifier: "MainTabBarController")
-        mainTabBarController.modalPresentationStyle = .overFullScreen
-        self.present(mainTabBarController, animated: true, completion: nil)
-    }
-}
